% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/trackeR.R
\name{append}
\alias{append}
\title{Generic function for appending data to existing files.}
\usage{
append(object, file, ...)
}
\arguments{
\item{object}{The object to be appended.}

\item{file}{The file to which \code{object} is to be appended.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for appending data to existing files.
}

