% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concentrationProfile.R, R/distributionProfile.R
\name{append.conProfile}
\alias{append.conProfile}
\alias{append.distrProfile}
\title{Append training profiles.}
\usage{
\method{append}{conProfile}(object, file, ...)

\method{append}{distrProfile}(object, file, ...)
}
\arguments{
\item{object}{The object to be appended.}

\item{file}{The file to which \code{object} is to be appended.}

\item{...}{Currently not used.}
}
\description{
Generic function for appending training distribution or
concentration profiles to existing files.
}

