% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{append_dynamics}
\alias{append_dynamics}
\title{Returns a dataframe with extra columns with dynamical information}
\usage{
append_dynamics(data_loc)
}
\arguments{
\item{data_loc}{The clean data from a given location}
}
\value{
A data frame including the speeds and accelerations
}
\description{
Returns a dataframe with extra columns with dynamical information
}
\seealso{
\code{\link{speed}, \link{accel}}
}
