% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{append_exp_info}
\alias{append_exp_info}
\title{Returns a dataframe with extra columns indicating experimental information}
\usage{
append_exp_info(data_loc, file_name)
}
\arguments{
\item{data_loc}{The clean data from a given location}

\item{file_name}{The original name of the file}
}
\value{
A data frame including the cosmnr
}
\description{
Returns a dataframe with extra columns indicating experimental information
}
