% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{append_light_interval}
\alias{append_light_interval}
\title{Returns a dataframe with extra columns containing the light interval, and light as a dummy variable}
\usage{
append_light_interval(data_loc)
}
\arguments{
\item{data_loc}{The clean data from a given location}
}
\value{
A data frame including the timebin
}
\description{
Returns a dataframe with extra columns containing the light interval, and light as a dummy variable
}
