% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{append_time_bins}
\alias{append_time_bins}
\title{Returns a dataframe with an extra column containing the timebin}
\usage{
append_time_bins(data_loc)
}
\arguments{
\item{data_loc}{The clean data from a given location}
}
\value{
A data frame including the timebin
}
\description{
Returns a dataframe with an extra column containing the timebin
}
