% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_assembly.R
\name{assemble_data}
\alias{assemble_data}
\title{Assemble all measurements, files, and experiments}
\usage{
assemble_data(test_date, test_species, data_loc = "../data/")
}
\arguments{
\item{test_date}{The date of the experiment}

\item{test_species}{The name of the test species}

\item{#save}{Dummy variable indicating intermittent saving}
}
\value{
All the data for one experiment
}
\description{
Assemble all measurements, files, and experiments
}
