% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{calculate_plot_height}
\alias{calculate_plot_height}
\title{Calculate plot height for either time in zones or work capacity}
\usage{
calculate_plot_height(metrics)
}
\arguments{
\item{metrics}{A vector of metrics that will be plotted.}
}
\description{
Calculate plot height for either time in zones or work capacity
}
