% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{change_object_units}
\alias{change_object_units}
\title{Change units of variables.}
\usage{
change_object_units(data, input, object)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}

\item{input}{An object of class \code{reactivevalues}.}

\item{object}{A character of either 'summary' or 'object' to specify which objects' units to change.}
}
\description{
Change units of variables.
}
