% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R, R/trackeR.R,
%   R/zzz_aliases.R
\name{change_units}
\alias{change_units}
\alias{change_units}
\alias{changeUnits}
\title{Change units of variables.}
\usage{
change_units(object, variable, unit, ...)

change_units(object, variable, unit, ...)

changeUnits(object, variable, unit, ...)
}
\arguments{
\item{object}{A character of either 'summary' or 'object' to specify which objects' units to change.}

\item{variable}{The variable of which the units of measurement are changed.}

\item{unit}{The unit of measurement to which is changed.}

\item{...}{Arguments to be passed to methods.}

\item{data}{An object of class \code{reactivevalues}.}

\item{input}{An object of class \code{reactivevalues}.}

\item{object}{The object of which the units of measurement are changed.}
}
\description{
Change units of variables.

Generic function for changing the units of measurement.
}
