% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_zones.R
\name{change_units.trackeRdataZones}
\alias{change_units.trackeRdataZones}
\title{Change the units of the variables in an \code{trackeRdataZones} object}
\usage{
\method{change_units}{trackeRdataZones}(object, variable, unit, ...)
}
\arguments{
\item{object}{An object of class \code{trackeRdataZones}.}

\item{variable}{A vector of variables to be changed. Note, these are expected to be
concepts like 'speed' rather than variable names like 'avgSpeed' or 'avgSpeedMoving'.}

\item{unit}{A vector with the units, corresponding to variable.}

\item{...}{Currently not used.}
}
\description{
Change the units of the variables in an \code{trackeRdataZones} object
}
