% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{change_units.trackeRthresholds}
\alias{change_units.trackeRthresholds}
\title{Change the units of the variables in an \code{trackeRthresholds} object}
\usage{
\method{change_units}{trackeRthresholds}(object, variable, unit, sport,
  ...)
}
\arguments{
\item{object}{An object of class \code{trackeRthresholds}.}

\item{variable}{A vector of variables whose units are to be changed.}

\item{unit}{A vector with the units, corresponding to \code{variable}.}

\item{sport}{A vector of sports (amongst \code{'cycling'},
\code{'running'}, \code{'swimming'}) with each element
corresponding to \code{variable} and \code{unit}.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Change the units of the variables in an \code{trackeRthresholds} object
}
