% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{change_units.trackeRWprime}
\alias{change_units.trackeRWprime}
\title{Change the units of the variables in an \code{trackeRdata} object}
\usage{
\method{change_units}{trackeRWprime}(object, variable, unit, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{variable}{A vector of variables to be changed.}

\item{unit}{A vector with the units, corresponding to variable.}

\item{...}{Currently not used.}
}
\description{
Change the units of the variables in an \code{trackeRdata} object
}
