% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{changeUnits}
\alias{changeUnits}
\title{Generic function for changing the units of measurement.}
\usage{
changeUnits(object, variable, unit, ...)
}
\arguments{
\item{object}{The object of which the units of measurement are changed.}

\item{variable}{The variable of which the units of measurement are changed.}

\item{unit}{The unit of measurement to which is changed.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for changing the units of measurement.
}

