% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{classify_sessions_by_sport}
\alias{classify_sessions_by_sport}
\title{Classify sessions by sport using the KNN model and 'sport_classification_train' dataset as a training set}
\usage{
classify_sessions_by_sport(data)
}
\description{
Classify sessions by sport using the KNN model and 'sport_classification_train' dataset as a training set
}
