% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{collect_units}
\alias{collect_units}
\title{Collect units from the result of \code{\link{generate_units}}}
\usage{
collect_units(object, unit_reference_sport = NULL)
}
\arguments{
\item{object}{a \code{data.frame}, as returned by
\code{\link{generate_units}}}

\item{unit_reference_sport}{The sport to inherit units from
(default is taken to be the most frequent sport in
\code{object}).}
}
\description{
Collects the units from the results of \code{\link{generate_units}}
according to a \code{unit_reference_sport}
}
