% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationProfile.R, R/trackeR.R
\name{concentration_profile.distrProfile}
\alias{concentration_profile.distrProfile}
\alias{conProfile}
\alias{concentration_profile.trackeRdata}
\alias{concentration_profile}
\title{Generate training concentration profiles.}
\usage{
\method{concentration_profile}{distrProfile}(object, session = NULL,
  what = NULL, ...)

\method{concentration_profile}{trackeRdata}(object, session = NULL,
  what = NULL, limits = NULL, parallel = FALSE,
  unit_reference_sport = NULL, scale = FALSE)

concentration_profile(object, session = NULL, what = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile} as returned by
\code{\link{distributionProfile}}.}

\item{session}{A numeric vector of the sessions to be used,
defaults to all sessions.}

\item{what}{The variables for which the concentration profiles
should be computed. Defaults (\code{NULL}) to all variables in
\code{object}.}

\item{...}{Currently not used.}

\item{parallel}{Logical. Should computation be carried out in
parallel? Default is \code{FALSE}.}
}
\value{
An object of class \code{conProfile}.

Object:

A named list with one or more components, corresponding to the
value of \code{what}. Each component is a matrix of dimension
\code{g} times \code{n}, where \code{g} is the length of the grids
set in \code{grid} (or 200 if \code{grid = NULL}) and \code{n} is
the number of sessions requested in the \code{session} argument.

Attributes:

Each \code{conProfile} object has the following attributes:

\itemize{

\item \code{sport}: the sports corresponding to the columns of each
list component

\item \code{session_times}: the session start and end times
correspoding to the columns of each list component

\item \code{unit_reference_sport}: the sport where the units have
been inherited from

\item \code{operations}: a list with the operations that have been
applied to the object. See \code{\link{get_operations.distrProfile}}

\item \code{units}: an object listing the units used for the
calculation of distribution profiles. These is the output of
\code{\link{get_units}} on the corresponding
\code{\link{trackeRdata}} object, after inheriting units from
\code{unit_reference_sport}.

}
}
\description{
Generate training concentration profiles.
}
\examples{
data('run', package = 'trackeR')
dProfile <- distributionProfile(run, what = 'speed', grid = seq(0, 12.5, by = 0.05))
cProfile <- concentrationProfile(dProfile)
plot(cProfile, smooth = FALSE)
plot(cProfile)
}
\references{
Kosmidis, I., and Passfield, L. (2015). Linking the Performance of
Endurance Runners to Training and Physiological Effects via
Multi-Resolution Elastic Net. \emph{ArXiv e-print}
arXiv:1506.01388.

Frick, H., Kosmidis, I. (2017). trackeR: Infrastructure for Running
and Cycling Data from GPS-Enabled Tracking Devices in
R. \emph{Journal of Statistical Software}, \bold{82}(7),
1--29. doi:10.18637/jss.v082.i07
}
