% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_option_box}
\alias{create_option_box}
\title{Create a return button from selected workouts plot}
\usage{
create_option_box(sport_options)
}
\arguments{
\item{sport_options}{A vector of sports identified from the uploaded sessions.}
}
\description{
Create a return button from selected workouts plot
}
