% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_profiles_box}
\alias{create_profiles_box}
\title{Create concentration profile plot UI.}
\usage{
create_profiles_box(title, inputId, label, plotId, choices)
}
\description{
Create concentration profile plot UI.
}
