% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_selected_workout_plot}
\alias{create_selected_workout_plot}
\title{Create selected_workouts plot}
\usage{
create_selected_workout_plot(id, collapsed = FALSE)
}
\arguments{
\item{id}{A character. The ID of the plot.}

\item{collapsed}{A logical. Whether or not the UI box should be collapsed.}
}
\description{
Create selected_workouts plot
}
