% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_summary_timeline_boxes}
\alias{create_summary_timeline_boxes}
\title{Create a summary and timeline boxes}
\usage{
create_summary_timeline_boxes()
}
\description{
Create a summary and timeline boxes
}
