% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_work_capacity_plot}
\alias{create_work_capacity_plot}
\title{Create work capacity plot}
\usage{
create_work_capacity_plot(id, collapsed = FALSE)
}
\arguments{
\item{id}{A character. The ID of the plot.}

\item{collapsed}{A logical. Whether or not the UI box should be collapsed.}
}
\description{
Create work capacity plot
}
