% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_workout_plots}
\alias{create_workout_plots}
\title{Create workout plots}
\usage{
create_workout_plots(feature)
}
\arguments{
\item{feature}{A character. The metric that is plotted, selected from \code{\link{choices}}.}
}
\description{
Create workout plots
}
