% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_zones_box}
\alias{create_zones_box}
\title{Create time in zones plot UI.}
\usage{
create_zones_box(inputId, plotId, choices)
}
\arguments{
\item{inputId}{A character. The ID of the user input for the metrics that should be plotted.}

\item{plotId}{A character. The ID of the plot.}

\item{choices}{A vector of the metrics a user can select to be plotted, selected from \code{\link{metrics}}.}
}
\description{
Create time in zones plot UI.
}
