% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{decreasingSmoother}
\alias{decreasingSmoother}
\title{Smooth a decreasing function.}
\usage{
decreasingSmoother(x, y, k = 30, len = NULL, sp = NULL, fam = "poisson")
}
\arguments{
\item{x}{The regressor passed on to the \code{formula} argument of \code{\link[scam]{scam}}.}

\item{y}{The response passed on to the \code{formula} argument of \code{\link[scam]{scam}}.}

\item{k}{Number of knots.}

\item{len}{If \code{NULL}, the default, \code{x} is used for prediction. Otherwise,
prediction is done over the range of \code{x} with \code{len} equidistant points.}

\item{sp}{A vector of smoothing parameters passed on to \code{\link[scam]{scam}}.}

\item{fam}{A family object passed on to the \code{family} argument of \code{\link[scam]{scam}}.}
}
\description{
This smoother ensures a positive response (Poisson) that is a monotone decreasing function of x.
}
