% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeSpeeds.R
\name{distance2speed}
\alias{distance2speed}
\title{Convert distance to speed.}
\usage{
distance2speed(distance, time)
}
\arguments{
\item{distance}{Distance in meters.}

\item{time}{Time.}
}
\value{
Speed in meters per second.
}
\description{
Convert distance to speed.
}

