% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R, R/zzz_aliases.R
\name{distribution_profile}
\alias{distribution_profile}
\alias{distributionProfile}
\title{Generate training distribution profiles.}
\usage{
distribution_profile(object, session = NULL, what = c("speed",
  "heart_rate"), grid = NULL, parallel = FALSE,
  unit_reference_sport = NULL)

distributionProfile(object, session = NULL, what = c("speed", "heart_rate"),
  grid = NULL, parallel = FALSE, unit_reference_sport = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be used,
defaults to all sessions.}

\item{what}{The variables for which the distribution profiles
should be generated.}

\item{grid}{A named list containing the grid values for the
variables in \code{what}. If \code{NULL} (default) the grids
for the variables in \code{what} are inferred from
\code{object}.}

\item{parallel}{Logical. Should computation be carried out in
parallel? Default is \code{FALSE}.}
}
\value{
An object of class \code{distrProfile}.
}
\description{
Generate training distribution profiles.
}
\examples{
data('run', package = 'trackeR')
dProfile <- distributionProfile(run, what = c("speed", "cadence"), grid = seq(0, 12.5, by = 0.05))
plot(dProfile, smooth = FALSE)
}
\references{
Kosmidis, I., and Passfield, L. (2015). Linking the Performance of
Endurance Runners to Training and Physiological Effects via
Multi-Resolution Elastic Net. \emph{ArXiv e-print}
arXiv:1506.01388.

Frick, H., Kosmidis, I. (2017). trackeR: Infrastructure for Running
and Cycling Data from GPS-Enabled Tracking Devices in
R. \emph{Journal of Statistical Software}, \bold{82}(7),
1--29. doi:10.18637/jss.v082.i07
}
