% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{download_handler}
\alias{download_handler}
\title{Download handler}
\usage{
download_handler(data)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}
}
\description{
Download handler
}
