% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filter_data}
\alias{filter_data}
\title{Returns a dataframe with rows filtered}
\usage{
filter_data(input_data, filter_location)
}
\arguments{
\item{input_data}{The clean data from all locations}

\item{filter_location}{The location on which the data will be filtered}
}
\value{
A data frame including only the rows belonging to one location
}
\description{
Returns a dataframe with rows filtered
}
