% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_unit_reference_sport}
\alias{find_unit_reference_sport}
\title{Find the most frequent sport in an \code{object}}
\usage{
find_unit_reference_sport(object)
}
\arguments{
\item{object}{any object with a \code{\link{get_sport}} method
implemented (run \code{methods(get_sport)}).}
}
\description{
Find the most frequent sport in an \code{object}
}
