% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackeRdata_plot.R
\name{fortify.trackeRdata}
\alias{fortify.trackeRdata}
\title{Fortify a trackeRdata object for plotting with ggplot2.}
\usage{
\method{fortify}{trackeRdata}(model, data, melt = FALSE, ...)
}
\arguments{
\item{model}{The \code{\link{trackeRdata}} object.}

\item{data}{Ignored.}

\item{melt}{Logical. Should the data be melted into long format
instead of the default wide format?}

\item{...}{Ignored.}
}
\description{
Fortify a trackeRdata object for plotting with ggplot2.
}

