% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/trackeRdata_summary.R
\name{fortify.trackeRdataSummary}
\alias{fortify.trackeRdataSummary}
\title{Fortify a trackeRdataSummary object for plotting with ggplot2.}
\usage{
\method{fortify}{trackeRdataSummary}(model, data, melt = FALSE, ...)
}
\arguments{
\item{model}{The \code{\link{trackeRdata}} object.}

\item{data}{Ignored.}

\item{melt}{Logical. Should the data be melted into long format
instead of the default wide format?}

\item{...}{Currently not used.}
}
\description{
Fortify a trackeRdataSummary object for plotting with ggplot2.
}

