% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wprime.R
\name{fortify.trackeRWprime}
\alias{fortify.trackeRWprime}
\title{Fortify a trackeRWprime object for plotting with ggplot2.}
\usage{
\method{fortify}{trackeRWprime}(model, data, melt = FALSE, ...)
}
\arguments{
\item{model}{The \code{trackeRWprime} object as returned by \code{\link{Wprime}}.}

\item{data}{Ignored.}

\item{melt}{Logical. Should the data be melted into long format
instead of the default wide format?}

\item{...}{Ignored.}
}
\description{
Fortify a trackeRWprime object for plotting with ggplot2.
}
