% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R, R/zzz_aliases.R
\name{generate_base_units}
\alias{generate_base_units}
\alias{generateBaseUnits}
\title{Generate base units}
\usage{
generate_base_units(...)

generateBaseUnits(...)
}
\arguments{
\item{...}{Currently not used}
}
\description{
Generate base units
}
