% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_modules.R
\name{generate_graph_data}
\alias{generate_graph_data}
\title{Generate a formatted data frame for \code{plot_workouts()} for a given variable.}
\usage{
generate_graph_data(x, what)
}
\arguments{
\item{x}{An object of class \code{trackeRdataSummary}.}

\item{what}{A character of a variable name to be plotted (e.g. "avgHeartRate").}
}
\description{
Generate a formatted data frame for \code{plot_workouts()} for a given variable.
}
