% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{generate_objects}
\alias{generate_objects}
\title{Classify sessions by sport, process dataset, generate download handler,
generate selected sessions object, update what metrics are available 
to plot and other minor actions.}
\usage{
generate_objects(data, output, session, choices)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}

\item{output}{A shiny object.}

\item{session}{A shiny object.}

\item{choices}{A vector. See \code{\link{choices}}.}
}
\description{
Classify sessions by sport, process dataset, generate download handler,
generate selected sessions object, update what metrics are available 
to plot and other minor actions.
}
