% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R, R/zzz_aliases.R
\name{generate_thresholds}
\alias{generate_thresholds}
\alias{generateDefaultThresholds}
\title{Generate default thresholds}
\usage{
generate_thresholds(variable, lower, upper, sport, ...)

generateDefaultThresholds(variable, lower, upper, sport, ...)
}
\arguments{
\item{...}{Currently not used.}
}
\description{
Generate default thresholds
}
