% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{generateBaseUnits}
\alias{generateBaseUnits}
\title{Generate base units.}
\usage{
generateBaseUnits(variable, unit, sport, ...)
}
\arguments{
\item{...}{Currently not used.}
}
\description{
Generate base units.
}
