% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{get_javascript}
\alias{get_javascript}
\title{Get Javascript code for reseting page and collapsing boxes}
\usage{
get_javascript()
}
\description{
Get Javascript code for reseting page and collapsing boxes
}
