% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationProfile.R
\name{get_operations.conProfile}
\alias{get_operations.conProfile}
\title{Get the operation settings of an \code{conProfile} object}
\usage{
\method{get_operations}{conProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{conProfile} as returned by \code{\link{concentrationProfile}}.}

\item{...}{Currently not used.}
}
\description{
Get the operation settings of an \code{conProfile} object
}
