% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_operations.distrProfile}
\alias{get_operations.distrProfile}
\title{Get the operation settings of an \code{distrProfile} object}
\usage{
\method{get_operations}{distrProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile} as returned by \code{\link{distributionProfile}}.}

\item{...}{Currently not used.}
}
\description{
Get the operation settings of an \code{distrProfile} object
}
