% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/zzz_aliases.R
\name{get_operations}
\alias{get_operations}
\alias{getOperations}
\title{Generic function for retrieving the operation settings.}
\usage{
get_operations(object, ...)

getOperations(object, ...)
}
\arguments{
\item{object}{The object of which the units of measurement are retrieved.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for retrieving the operation settings.
}
