% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{get_operations.trackeRdata}
\alias{get_operations.trackeRdata}
\title{Get the operation settings of an \code{trackeRdata} object}
\usage{
\method{get_operations}{trackeRdata}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{...}{Currently not used.}
}
\description{
Get the operation settings of an \code{trackeRdata} object
}
