% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{get_profile}
\alias{get_profile}
\title{Extract specific distribution and concentration profiles from
\code{\link{distrProfile}} and \code{\link{conProfile}} objects}
\usage{
get_profile(object, session = NULL, what = c("speed", "heart_rate"), ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile} as returned by
\code{\link{distributionProfile}}.}

\item{session}{A numeric vector of the sessions to selected.
Defaults to all sessions.}

\item{what}{A character version of the variables to be
selected. Defaults to \code{c('speed', 'heart_rate')}.}
}
\description{
Extract specific distribution and concentration profiles from
\code{\link{distrProfile}} and \code{\link{conProfile}} objects
}
