% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R, R/zzz_aliases.R
\name{get_resting_periods}
\alias{get_resting_periods}
\alias{restingPeriods}
\title{Extract resting period characteristics}
\usage{
get_resting_periods(times, session_threshold)

restingPeriods(times, session_threshold)
}
\arguments{
\item{times}{Timestamps.}

\item{session_threshold}{The threshold in hours for the time
difference between consecutive timestamps above which they are
considered to belong to different training sessions.}
}
\value{
A list containing a dataframe with start, end, and duration
    for each session and the resting time between sessions, named
    'sessions' and 'restingTime', respectively.
}
\description{
Extract resting period characteristics
}
