% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{get_selected_units}
\alias{get_selected_units}
\title{Get units of measurement for a given feature}
\usage{
get_selected_units(feature, data)
}
\arguments{
\item{feature}{A character for the feature whose units we want to access, for example 'altitude', 'distance',...}

\item{data}{An object of class \code{reactivevalues}.}
}
\description{
Get units of measurement for a given feature
}
