% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wprime.R, R/trackeR.R, R/trackeRdata.R,
%   R/trackeRdata_summary.R
\name{get_sport.trackeRWprime}
\alias{get_sport.trackeRWprime}
\alias{get_sport}
\alias{get_sport.trackeRdata}
\alias{get_sport.trackeRdataSummary}
\title{Generic function for extracting sports}
\usage{
\method{get_sport}{trackeRWprime}(object, ...)

get_sport(object, ...)

\method{get_sport}{trackeRdata}(object, ...)

\method{get_sport}{trackeRdataSummary}(object, ...)
}
\arguments{
\item{object}{The object from which to extract sports.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for extracting sports
}
