% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_units.conProfile}
\alias{get_units.conProfile}
\title{Get the units of the variables in an \code{conProfile} object}
\usage{
\method{get_units}{conProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{conProfile}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{conProfile} object
}
