% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_units.distrProfile}
\alias{get_units.distrProfile}
\title{Get the units of the variables in an \code{distrProfile} object.}
\usage{
\method{get_units}{distrProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{distrProfile} object.
}
