% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/zzz_aliases.R
\name{get_units}
\alias{get_units}
\alias{getUnits}
\title{Generic function for extracting the units of measurement.}
\usage{
get_units(object, ...)

getUnits(object, ...)
}
\arguments{
\item{object}{The object of which the units of measurement are retrieved.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for extracting the units of measurement.
}
