% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{get_units.trackeRdata}
\alias{get_units.trackeRdata}
\title{Get the units of the variables in an \code{trackeRdata} object}
\usage{
\method{get_units}{trackeRdata}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRdata} object
}
