% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_zones.R
\name{get_units.trackeRdataZones}
\alias{get_units.trackeRdataZones}
\title{Get the units of the variables in an \code{trackeRdataZones} object}
\usage{
\method{get_units}{trackeRdataZones}(object, ...)
}
\arguments{
\item{object}{An object of class \code{trackeRdataZones}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRdataZones} object
}
