% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_units.trackeRfpca}
\alias{get_units.trackeRfpca}
\title{Get the units of the variables in an \code{trackeRfpca} object}
\usage{
\method{get_units}{trackeRfpca}(object, ...)
}
\arguments{
\item{object}{An object of class \code{trackeRfpca}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRfpca} object
}
