% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{get_units.trackeRthresholds}
\alias{get_units.trackeRthresholds}
\title{Get the units of the variables in an \code{trackeRthresholds} object}
\usage{
\method{get_units}{trackeRthresholds}(object, ...)
}
\arguments{
\item{object}{An object of class \code{trackeRthresholds}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRthresholds} object
}
