% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/units.R
\name{getOperations.conProfile}
\alias{getOperations.conProfile}
\title{Get the operation settings of an \code{conProfile} object.}
\usage{
\method{getOperations}{conProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{conProfile} as returned by \code{\link{concentrationProfile}}.}

\item{...}{Currently not used.}
}
\description{
Get the operation settings of an \code{conProfile} object.
}

