% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/units.R
\name{getOperations.distrProfile}
\alias{getOperations.distrProfile}
\title{Get the operation settings of an \code{distrProfile} object.}
\usage{
\method{getOperations}{distrProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile} as returned by \code{\link{distributionProfile}}.}

\item{...}{Currently not used.}
}
\description{
Get the operation settings of an \code{distrProfile} object.
}

