% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackeR.R
\name{getOperations}
\alias{getOperations}
\title{Generic function for retrieving the operation settings.}
\usage{
getOperations(object, ...)
}
\arguments{
\item{object}{The object of which the units of measurement are retrieved.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for retrieving the operation settings.
}

