% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/units.R
\name{getOperations.trackeRdata}
\alias{getOperations.trackeRdata}
\title{Get the operation settings of an \code{trackeRdata} object.}
\usage{
\method{getOperations}{trackeRdata}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{...}{Currently not used.}
}
\description{
Get the operation settings of an \code{trackeRdata} object.
}

