% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getUnits.conProfile}
\alias{getUnits.conProfile}
\title{Get the units of the variables in an \code{conProfile} object.}
\usage{
\method{getUnits}{conProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{conProfile}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{conProfile} object.
}
