% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/units.R
\name{getUnits.distrProfile}
\alias{getUnits.distrProfile}
\title{Get the units of the variables in an \code{distrProfile} object.}
\usage{
\method{getUnits}{distrProfile}(object, ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{distrProfile} object.
}

