% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{getUnits}
\alias{getUnits}
\title{Generic function for retrieving the units of measurement.}
\usage{
getUnits(object, ...)
}
\arguments{
\item{object}{The object of which the units of measurement are retrieved.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for retrieving the units of measurement.
}

