% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/units.R
\name{getUnits.trackeRdata}
\alias{getUnits.trackeRdata}
\title{Get the units of the variables in an \code{trackeRdata} object.}
\usage{
\method{getUnits}{trackeRdata}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRdata} object.
}

