% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getUnits.trackeRdataSummary}
\alias{getUnits.trackeRdataSummary}
\title{Get the units of the variables in an \code{trackeRdataSummary} object.}
\usage{
\method{getUnits}{trackeRdataSummary}(object, ...)
}
\arguments{
\item{object}{An object of class \code{trackeRdataSummary}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRdataSummary} object.
}
