% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{getUnits.trackeRWprime}
\alias{getUnits.trackeRWprime}
\title{Get the units of the variables in an \code{trackeRWprime} object.}
\usage{
\method{getUnits}{trackeRWprime}(object, ...)
}
\arguments{
\item{object}{An object of class \code{trackeRWprime}.}

\item{...}{Currently not used.}
}
\description{
Get the units of the variables in an \code{trackeRWprime} object.
}
