% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_shiny.R
\name{lab_sum}
\alias{lab_sum}
\title{Generate a character of formatted units, either only the unit (e.g "[bpm]") or whole text (e.g. "Heart Rate \n [bpm]").}
\usage{
lab_sum(feature, data, whole_text = TRUE, transform_feature = TRUE)
}
\arguments{
\item{feature}{A character representing the feature whose units we want to generate.}

\item{data}{An object of class \code{trackeRdataSummary} or \code{trackeRdata}.}

\item{whole_text}{Generate only unit (e.g "[bpm]") or whole text (e.g. "Heart Rate \n [bpm]").}

\item{transform_feature}{If TRUE, epected format of \code{feature} is such as "avgCadence", "avgPower". If FALSE, expected format is "pace", "cadence", "heart.rate" or "altitude".}
}
\description{
Generate a character of formatted units, either only the unit (e.g "[bpm]") or whole text (e.g. "Heart Rate \n [bpm]").
}
