% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{nsessions}
\alias{nsessions}
\title{Generic function for calculating number of sessions.}
\usage{
nsessions(object, ...)
}
\arguments{
\item{object}{The object for which to calculate the number of sessions.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for calculating number of sessions.
}
