% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_zones.R
\name{plot.trackeRdataZones}
\alias{plot.trackeRdataZones}
\title{Plot training zones.}
\usage{
plot.trackeRdataZones(x, percent = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{trackeRdataZones} as returned by \code{\link{zones}}.}

\item{percent}{Logical. Should the relative or absolute times spent training in the different zones be plotted?}

\item{...}{Currently not used.}
}
\description{
Plot training zones.
}
\examples{
data('run', package = 'trackeR')
runZones <- zones(run, what = 'speed', breaks = c(0, 2:6, 12.5))
plot(runZones, percent = FALSE)
}
