% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Wprime.R
\name{plot.trackeRWprime}
\alias{plot.trackeRWprime}
\title{Plot W'.}
\usage{
\method{plot}{trackeRWprime}(x, session = NULL, dates = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{trackeRWprime} as returned by \code{\link{Wprime}}.}

\item{session}{A numeric vector of the sessions to be plotted, defaults to all sessions.}

\item{dates}{Logical. Should the date of the session be used in the panel header?}

\item{...}{Currently not used.}
}
\description{
Plot W'.
}
\examples{
data("runs", package = "trackeR")
wexp <- Wprime(runs, session = 1:3, cp = 4, version = "2012")
plot(wexp, session = 1:2)
}

