% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap2}
\alias{plot_heatmap2}
\title{Plot more sophisticated heatmap}
\usage{
plot_heatmap2(
  input_acute,
  input_chronic,
  chemical,
  species,
  exp_dur,
  timebin,
  test_location
)
}
\arguments{
\item{input_acute}{Dataframe of acute exposure}

\item{input_chronic}{Dataframe of chronic exposure}

\item{chemical}{Name of chemical}

\item{species}{Name of test species}

\item{exp_dur}{Duration of the experiment in seconds}

\item{timebin}{Size of the timebin in seconds}
}
\value{
The plot
}
\description{
Plot more sophisticated heatmap
}
