% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_positions}
\alias{plot_positions}
\title{Plot positions}
\usage{
plot_positions(data_loc, type = "scatter")
}
\arguments{
\item{data_loc}{Clean dataframe corresponding to a location}

\item{type}{Type of plot (available types: 'scatter', 'density')}
}
\value{
The plot
}
\description{
Plot positions
}
