% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{A timeline plot for workouts.}
\usage{
plot_timeline(sumX, lims = NULL, shiny = TRUE, plotly = TRUE)
}
\arguments{
\item{sumX}{An object of class \code{trackeRdataSummary}.}

\item{shiny}{Logical. Whether plots are in a shiny environment.}

\item{plotly}{Logical. Return plotly plots or standard TrackeR plots}
}
\description{
A timeline plot for workouts.
}
