% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_velocity.R
\name{plot_velocity}
\alias{plot_velocity}
\title{Plot positions}
\usage{
plot_velocity(test_date, chemical, timebin, exp_dur)
}
\arguments{
\item{test_date}{Date of the experiment}

\item{chemical}{Name of chemical}

\item{timebin}{Size of the timebin}

\item{exp_dur}{Duration of the experiment in seconds}
}
\value{
The plot
}
\description{
Plot positions
}
