% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_velocity2.R
\name{plot_velocity2}
\alias{plot_velocity2}
\title{Plot positions}
\usage{
plot_velocity2(input_data, group_id, compare_id)
}
\arguments{
\item{group_id}{Columnname used for grouping}

\item{compare_id}{Columnname used for comparing}

\item{input1}{First dataframe with behavioural data}

\item{input2}{Second dataframe with behavioural data}

\item{figure_name}{Name of the output figure}
}
\value{
The plot
}
\description{
Plot positions
}
