% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{plot_work_capacities}
\alias{plot_work_capacities}
\title{#' Calculate plot height for work capacity
calculate_plot_height_work_capacity <- function(name, data) {
  if (name != 'work_capacity') {
    "250px"
  } else {
    sports <- unique(sport(data$object[data$selectedSessions]))
    # Work capacity only for running and cycling
    sports <- intersect(c('running', 'cycling'), sports)
    paste0(250 * length(sports), "px")
  }
}
Plot work capacities for each sport}
\usage{
plot_work_capacities(x, session, cp)
}
\arguments{
\item{x}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be used, defaults to all sessions.}

\item{cp}{A numeric. Critical power/speed, i.e., the power/speed which can be maintained for longer period of time.}
}
\description{
#' Calculate plot height for work capacity
calculate_plot_height_work_capacity <- function(name, data) {
  if (name != 'work_capacity') {
    "250px"
  } else {
    sports <- unique(sport(data$object[data$selectedSessions]))
    # Work capacity only for running and cycling
    sports <- intersect(c('running', 'cycling'), sports)
    paste0(250 * length(sports), "px")
  }
}
Plot work capacities for each sport
}
