% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_work_capacity.R
\name{plot_work_capacity}
\alias{plot_work_capacity}
\title{Plot the work capacity W' (w prime).}
\usage{
plot_work_capacity(run_data, session)
}
\arguments{
\item{run_data}{An object of class \code{trackeRdata}.}

\item{session}{A vector of selected sessions.}
}
\description{
Plot the work capacity W' (w prime).
}
