% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_shiny.R
\name{plot_workouts}
\alias{plot_workouts}
\title{A time-series plotly graph for a given variable (e.g. "avgHeartRate").}
\usage{
plot_workouts(dat, feature, name, units)
}
\arguments{
\item{dat}{A dataframe generated using \code{generate_graph_data()}.}

\item{feature}{A character for y-axis title generated using \code{lab_sum()}.}

\item{name}{A character for the feature to be plotted (e.g. "avgHeartRate").}

\item{units}{A character of the unit of measurement for the given variable (e.g. "[bmp]") generated using \code{lab_sum()}.}
}
\description{
A time-series plotly graph for a given variable (e.g. "avgHeartRate").
}
