% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zones_shiny.R
\name{plot_zones}
\alias{plot_zones}
\title{Plot training zones}
\usage{
plot_zones(x, session, what = c("speed"))
}
\arguments{
\item{x}{An object of class \code{trackeRdata} as returned by \code{\link{readDirectory_shiny}}.}

\item{session}{A vector of selected sessions.}

\item{what}{A vector of variable names to be plotted.}
}
\description{
Plot training zones
}
