% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_plot.R
\name{plotRoute2}
\alias{plotRoute2}
\title{Plot routes for training sessions.}
\usage{
plotRoute2(x, session = 1, speed = TRUE, threshold = TRUE, ...)
}
\arguments{
\item{x}{A object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be plotted. Defaults to
the first session, all sessions can be plotted by \code{session = NULL}.}

\item{speed}{Logical. Should the trace be colored according to speed?}

\item{threshold}{Logical. Should thresholds be applied?}

\item{...}{Additional arguments passed on to \code{\link{threshold}}.}
}
\description{
Plot the route ran/cycled during training on an interactive map.
Internet connection is required to download the background map.
}
\examples{
\dontrun{
data(runs, package = "trackeR")
plotRoute2(runs, session = 22:23)
}
}

