% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_plot.R
\name{prettifyUnit}
\alias{prettifyUnit}
\alias{prettifyUnits}
\title{Returns 'pretty' units for use for plotting or printing}
\usage{
prettifyUnit(unit)

prettifyUnits(unit)
}
\arguments{
\item{unit}{a unit as recorded in the \code{\link{data.frame}}
generated by \code{\link{generate_units}}.}
}
\description{
Returns 'pretty' units for use for plotting or printing
}
\details{
\code{prettifyUnits} is the vectorized version of \code{prettifyUnit}
}
\examples{
prettifyUnit("m_per_s")
prettifyUnit("rev_per_min")
prettifyUnits(c("rev_per_min", "ft_per_min"))

}
