% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_summary.R
\name{print.trackeRdataSummary}
\alias{print.trackeRdataSummary}
\title{Print method for session summaries.}
\usage{
\method{print}{trackeRdataSummary}(x, ..., digits = 2)
}
\arguments{
\item{x}{An object of class \code{trackeRdataSummary}.}

\item{...}{Not used, for compatibility with generic summary method only.}

\item{digits}{Number of digits to be printed.}
}
\description{
Print method for session summaries.
}

