% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{process_dataset}
\alias{process_dataset}
\title{Process \code{trackeRdata} object by: setting thresholds to remove wrong values, change units, set a moving threshold and test which variables contain data}
\usage{
process_dataset(data)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}
}
\description{
Process \code{trackeRdata} object by: setting thresholds to remove wrong values, change units, set a moving threshold and test which variables contain data
}
