% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/read.R
\name{readContainer}
\alias{readContainer}
\title{Read a GPS container file.}
\usage{
readContainer(file, type = c("tcx", "db3"), table = "gps_data",
  timezone = "", sessionThreshold = 2, fromDistances = NULL,
  speedunit = switch(type, tcx = "m_per_s", db3 = "km_per_h"),
  distanceunit = switch(type, tcx = "m", db3 = "km"), cycling = FALSE,
  lgap = 30, lskip = 5, m = 11, mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{file}{The path to the file.}

\item{type}{The type of the GPS container file. Supported so far are \code{tcx} and \code{db3}.}

\item{table}{The name of the table in the database if \code{type} is set to \code{db3},
ignored otherwise.}

\item{timezone}{The timezone of the observations as passed on to \code{\link[base]{as.POSIXct}}.}

\item{sessionThreshold}{The threshold in hours for the time
difference between consecutive timestamps above which they are
considered to belong to different training sessions.}

\item{fromDistances}{Logical. Should the speeds be calculated from the distance recordings
instead of taken from the speed recordings directly. Defaults to \code{TRUE} for \code{tcx}
files and to \code{FALSE} for \code{db3} files.}

\item{speedunit}{Character string indicating the measurement unit of the speeds in the container
file to be converted into meters per second. Default is \code{m_per_s} when \code{type} is
\code{tcx} and \code{km_per_h} when \code{type} is \code{db3}. See Details.}

\item{distanceunit}{Character string indicating the measurement unit of the distance in the container
file to be converted into meters. Default is \code{m} when \code{type} is
\code{tcx} and \code{km} when \code{type} is \code{db3}. See Details.}

\item{cycling}{Logical. Do the data stem from cycling instead of running? If so, the unit of
measurement for cadence is set to \code{rev_per_min} instead of \code{steps_per_min}.}

\item{lgap}{Time in seconds corresponding to the minimal sampling rate.}

\item{lskip}{Time in seconds between the last observation before a small break
and the first imputed speed or the last imputed speed and the first
observation after a small break.}

\item{m}{Number of imputed observations in each small break.}

\item{mc.cores}{Number of cores for parallel computing.}
}
\value{
An object of class \code{\link{trackeRdata}}.
}
\description{
Read a GPS container file.
}
\details{
Available options for \code{speedunit} currently are \code{km_per_h}, \code{m_per_s},
    \code{mi_per_h}, \code{ft_per_min} and \code{ft_per_s}.
    Available options for \code{distanceunit} currently are \code{km}, \code{m}, \code{mi} and
    \code{ft}.
}
\examples{
\dontrun{
filepath <- system.file("extdata", "2013-06-04-174137.TCX", package = "trackeR")
run <- readContainer(filepath, type = "tcx", timezone = "GMT")
}
}
\seealso{
\code{\link{trackeRdata}}, \code{\link{readTCX}}, \code{\link{readDB3}}
}

