% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readDirectory}
\alias{readDirectory}
\title{Read all supported container files from a supplied directory.}
\usage{
readDirectory(directory, aggregate = TRUE, table = "gps_data",
  timezone = "", sessionThreshold = 2, correctDistances = FALSE,
  country = NULL, mask = TRUE, fromDistances = NULL,
  speedunit = list(tcx = "m_per_s", db3 = "km_per_h"),
  distanceunit = list(tcx = "m", db3 = "km"), cycling = FALSE, lgap = 30,
  lskip = 5, m = 11, parallel = FALSE, mc.cores = getOption("mc.cores",
  2L), verbose = TRUE)
}
\arguments{
\item{directory}{The path to the directory.}

\item{aggregate}{Logical. Aggregate data from different files to the same session if observations are less then \code{sessionThreshold} hours apart? Alternatively, data from different files is stored in different sessions.}

\item{table}{The name of the table in the database for db3 files.}

\item{timezone}{The timezone of the observations as passed on to \code{\link[base]{as.POSIXct}}.}

\item{sessionThreshold}{The threshold in hours for the time
difference between consecutive timestamps above which they are
considered to belong to different training sessions.}

\item{correctDistances}{Logical. Should the distances be corrected for elevation?}

\item{country}{ISO3 country code for downloading altitude data. If \code{NULL}, country is derived from
longitude and latitude.}

\item{mask}{Logical. Passed on to \code{\link[raster]{getData}}. Should only the altitudes for the specified
\code{country} be extracted (\code{TRUE}) or also those for the neighboring countries (\code{FALSE})?}

\item{fromDistances}{Logical. Should the speeds be calculated from the distance recordings
instead of taken from the speed recordings directly. Defaults to \code{TRUE} for tcx
files and to \code{FALSE} for db3 files.}

\item{speedunit}{Character string indicating the measurement unit of the speeds in the container
file to be converted into meters per second. Default is \code{m_per_s} for tcx files and \code{km_per_h} for db3 files. See Details.}

\item{distanceunit}{Character string indicating the measurement unit of the distance in the container
file to be converted into meters. Default is \code{m} for tcx files and \code{km} for db3 files. See Details.}

\item{cycling}{Logical. Do the data stem from cycling instead of running? If so, the default unit of
measurement for cadence is set to \code{rev_per_min} instead of \code{steps_per_min} and power is
imputed with \code{0}, else with \code{NA}.}

\item{lgap}{Time in seconds corresponding to the minimal sampling rate.}

\item{lskip}{Time in seconds between the last observation before a small break
and the first imputed speed or the last imputed speed and the first
observation after a small break.}

\item{m}{Number of imputed observations in each small break.}

\item{parallel}{Logical. Should computation be carried out in parallel? (Not supported on Windows.)}

\item{mc.cores}{Number of cores for parallel computing.}

\item{verbose}{Logical. Should progress reports be printed?}
}
\value{
An object of class \code{\link{trackeRdata}}.
}
\description{
Read all supported container files from a supplied directory.
}
\details{
Available options for \code{speedunit} currently are \code{km_per_h}, \code{m_per_s},
    \code{mi_per_h}, \code{ft_per_min} and \code{ft_per_s}.
    Available options for \code{distanceunit} currently are \code{km}, \code{m}, \code{mi} and
    \code{ft}.
}
\seealso{
\code{\link{trackeRdata}}, \code{\link{readTCX}}, \code{\link{readDB3}}
}

