% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDirectory_shiny.R
\name{readDirectory_shiny}
\alias{readDirectory_shiny}
\title{Same function as \code{trackeR::readDirectory()}. Customized for shiny interface with a loading bar in the UI using \code{shinycssloaders} package.}
\usage{
readDirectory_shiny(input, output, session, directory, aggregate = TRUE,
  table = "gps_data", timezone = "", sessionThreshold = 2,
  correctDistances = FALSE, country = NULL, mask = TRUE,
  fromDistances = NULL, speedunit = list(tcx = "m_per_s", db3 = "km_per_h",
  json = "km_per_h"), distanceunit = list(tcx = "m", db3 = "km", json = "km"),
  cycling = FALSE, lgap = 30, lskip = 5, m = 11, silent = FALSE,
  parallel = FALSE, cores = getOption("mc.cores", 2L), verbose = TRUE)
}
\arguments{
\item{...}{same arguments as \code{trackeR::readDirectory()}.}
}
\description{
Same function as \code{trackeR::readDirectory()}. Customized for shiny interface with a loading bar in the UI using \code{shinycssloaders} package.
}
