% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryInput.R
\name{readDirectoryInput}
\alias{readDirectoryInput}
\title{Read the value of a directoryInput}
\usage{
readDirectoryInput(session, inputId)
}
\arguments{
\item{session}{The \code{session} object passed to function given to \code{shinyServer}.}

\item{inputId}{The id of the input object}
}
\description{
Read the value of a directoryInput
}
\details{
Reads the value of the text field associated with a \code{directoryInput}
object that stores the user selected directory path.
}
