% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_output.R
\name{render_summary_table}
\alias{render_summary_table}
\title{Render summary table}
\usage{
render_summary_table(data, input)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}

\item{input}{A shiny object with user input.}
}
\description{
Render summary table
}
