% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{ridges.distrProfile}
\alias{ridges.distrProfile}
\title{Ridgeline plots for \code{distrProfile} objects}
\usage{
\method{ridges}{distrProfile}(x, session = NULL, what = c("speed"),
  smooth = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{distrProfile} as returned by \code{\link{distributionProfile}}.}

\item{session}{A numeric vector of the sessions to be plotted, defaults to all sessions.}

\item{what}{Which variables should be plotted?}

\item{smooth}{Logical. Should unsmoothed profiles be smoothed before plotting?}

\item{...}{Further arguments to be passed to \code{\link{smootherControl.distrProfile}}.}
}
\description{
Ridgeline plots for \code{distrProfile} objects
}
\examples{
\dontrun{

data('runs', package = 'trackeR')
dProfile <- distributionProfile(runs, what = c("speed", "heart.rate"), auto_grid = TRUE)
ridges(dProfile)

}
}
