% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{ridges}
\alias{ridges}
\title{Generic function for ridgeline plots}
\usage{
ridges(x, ...)
}
\arguments{
\item{x}{An object of class \code{distrProfile} or
\code{conProfile}.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for ridgeline plots
}
