% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_plot.R
\name{ridges.trackeRdata}
\alias{ridges.trackeRdata}
\title{Ridgeline plots for \code{trackeRdata} objects}
\usage{
\method{ridges}{trackeRdata}(x, session = NULL, what = "speed", ...)
}
\arguments{
\item{session}{A numeric vector of the sessions to be used, defaults to all sessions.}

\item{what}{The variables for which the distribution profiles should be generated.}

\item{...}{Further arguments to be passed to \code{\link{smootherControl.distrProfile}}.}
}
\description{
Ridgeline plots for \code{trackeRdata} objects
}
\examples{

\dontrun{
data("runs", package = "trackeR")
ridges(runs)
}

}
