% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{runs}
\alias{runs}
\title{Training sessions.}
\format{A \code{\link{trackeRdata}} object containing 26 running training sessions.}
\usage{
runs
}
\description{
Training sessions.
}
\keyword{datasets}

