% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_utilities.R
\name{sanity_checks}
\alias{sanity_checks}
\title{Sanity checks for tracking data.}
\usage{
sanity_checks(dat, silent)
}
\arguments{
\item{dat}{Data set to be cleaned up.}

\item{silent}{Logical. Should warnings be generated if any of the
sanity checks on the data are triggered?}
}
\description{
Heart rate measurements of 0 are set to NA, assuming the athlete is alive.
Observations with missing or duplicated time stamps are removed.
}
