% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{sanityChecks}
\alias{sanityChecks}
\title{Sanity checks for tracking data.}
\usage{
sanityChecks(dat, silent)
}
\arguments{
\item{dat}{Data set to be clean up.}

\item{silent}{Logical. Should warnings be generated if any of the sanity checks on the data are triggered?}
}
\description{
Heart rate measurements of 0 are set to NA, assuming the athlete is alive.
Observations with missing or duplicated time stamps are removed.
}
