% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{scaled.distrProfile}
\alias{scaled.distrProfile}
\title{Scale the distribution profile relative to its maximum value}
\usage{
\method{scaled}{distrProfile}(object, session = NULL, what = c("speed",
  "heart_rate"), ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile} as returned by \code{\link{distributionProfile}}.}

\item{session}{A numeric vector of the sessions to be plotted, defaults to all sessions.}

\item{what}{Which variables should be scaled?}

\item{...}{Currently not used.}
}
\description{
Scale the distribution profile relative to its maximum value
}
