% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{scaled}
\alias{scaled}
\title{Generic function for scaling.}
\usage{
scaled(object, ...)
}
\arguments{
\item{...}{Arguments to be passed to methods.}

\item{x}{The object to be smoothed.}
}
\description{
Generic function for scaling.
}

