% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/trackeRdata.R,
%   R/trackeRdata_summary.R
\name{session_duration}
\alias{session_duration}
\alias{session_duration.trackeRdata}
\alias{session_duration.trackeRdataSummary}
\title{Generic function for calculating session durations}
\usage{
session_duration(object, session, ...)

\method{session_duration}{trackeRdata}(object, session = NULL, ...)

\method{session_duration}{trackeRdataSummary}(object, session = NULL, ...)
}
\arguments{
\item{object}{The object for which to calculate session durations.}

\item{session}{The sessions for which to extract sports.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for calculating session durations
}
\details{
The times units will be inherited from \code{object}.
}
