% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/trackeRdata.R,
%   R/trackeRdata_summary.R
\name{session_times}
\alias{session_times}
\alias{session_times.trackeRdata}
\alias{session_times.trackeRdataSummary}
\title{Generic function for calculating session times}
\usage{
session_times(object, session, ...)

\method{session_times}{trackeRdata}(object, session = NULL, ...)

\method{session_times}{trackeRdataSummary}(object, session = NULL, ...)
}
\arguments{
\item{object}{The object for which to calculate session start and end times.}

\item{session}{The sessions for which to extract sports.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for calculating session times
}
