% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_shiny.R
\name{shiny_plot_map}
\alias{shiny_plot_map}
\title{Plot all selected workouts on an interactive leaflet map.}
\usage{
shiny_plot_map(x, session, data_summary)
}
\arguments{
\item{x}{An object of class \code{trackeRdata}.}

\item{session}{A vector of selected sessions.}

\item{data_summary}{An object of class \code{trackeRdataSummary}.}
}
\description{
Plot all selected workouts on an interactive leaflet map.
}
