% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{show_change_unit_window}
\alias{show_change_unit_window}
\title{Generate a modal window where user can chage units of measurement.}
\usage{
show_change_unit_window(data)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}
}
\description{
Generate a modal window where user can chage units of measurement.
}
