% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackeR.R
\name{smoother}
\alias{smoother}
\title{Generic function for smoothing.}
\usage{
smoother(object, ...)
}
\arguments{
\item{object}{The object to be smoothed.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for smoothing.
}

