% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R, R/zzz_aliases.R
\name{smoother_control.distrProfile}
\alias{smoother_control.distrProfile}
\alias{smootherControl.distrProfile}
\title{Auxiliary function for
\code{\link{smoother.distrProfile}}. Typically used to construct a
control argument for \code{\link{smoother.distrProfile}}.}
\usage{
smoother_control.distrProfile(what = c("speed", "heart_rate"), k = 30,
  sp = NULL, parallel = FALSE)

smootherControl.distrProfile(what = c("speed", "heart_rate"), k = 30,
  sp = NULL, parallel = FALSE)
}
\arguments{
\item{what}{Vector of the names of the variables which should be
smoothed.}

\item{k}{Number of knots.}

\item{sp}{A vector of smoothing parameters passed on to \code{\link[scam]{scam}}.}

\item{parallel}{Logical. Should computation be carried out in
parallel?}
}
\description{
Auxiliary function for
\code{\link{smoother.distrProfile}}. Typically used to construct a
control argument for \code{\link{smoother.distrProfile}}.
}
