% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{smootherControl.distrProfile}
\alias{smootherControl.distrProfile}
\title{Auxiliary function for \code{\link{smoother.distrProfile}}. Typically used to construct
a control argument for \code{\link{smoother.distrProfile}}.}
\usage{
smootherControl.distrProfile(what = c("speed", "heart.rate"), k = 30,
  sp = NULL, parallel = FALSE, cl = NULL, cores = NULL)
}
\arguments{
\item{what}{Vector of the names of the variables which should be smoothed.}

\item{k}{Number of knots.}

\item{sp}{A vector of smoothing parameters passed on to \code{\link[scam]{scam}}.}

\item{parallel}{Logical. Should computation be carried out in parallel?}

\item{cl}{Passed on to \code{\link[doParallel]{registerDoParallel}}.}

\item{cores}{Number of cores for parallel computing.}
}
\description{
Auxiliary function for \code{\link{smoother.distrProfile}}. Typically used to construct
a control argument for \code{\link{smoother.distrProfile}}.
}

