% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{sort.trackeRdata}
\alias{sort.trackeRdata}
\title{Sort sessions in \code{\link{trackeRdata}} objects}
\usage{
\method{sort}{trackeRdata}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{A \code{trackeRdata} object.}

\item{decreasing}{Logical. Should the objects be sorted in
increasing or decreasing order?}

\item{...}{Currently not used.}
}
\description{
Sort the sessions \code{\link{trackeRdata}} objects into ascending
or descending order according to the first session timestamp.
}
