% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputeSpeeds.R
\name{speed2distance}
\alias{speed2distance}
\title{Convert speed to distance.}
\usage{
speed2distance(speed, time, cumulative = TRUE)
}
\arguments{
\item{speed}{Speed in meters per second.}

\item{time}{Time.}

\item{cumulative}{Logical. Should the cumulative distances be returned?}
}
\value{
Distance in meters.
}
\description{
Convert speed to distance.
}

