% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wprime.R, R/trackeR.R, R/trackeRdata.R,
%   R/trackeRdata_summary.R
\name{sport.trackeRWprime}
\alias{sport.trackeRWprime}
\alias{sport}
\alias{sport.trackeRdata}
\alias{sport.trackeRdataSummary}
\title{Generic function for extracting sports}
\usage{
\method{sport}{trackeRWprime}(object, ...)

sport(object, ...)

\method{sport}{trackeRdata}(object, ...)

\method{sport}{trackeRdataSummary}(object, ...)
}
\arguments{
\item{object}{The object from which to extract sports.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for extracting sports
}
