% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\docType{data}
\name{sports_options}
\alias{sports_options}
\title{Currently available sports in the trackeRdashboard.}
\format{An object of class \code{character} of length 3.}
\usage{
sports_options
}
\description{
Currently available sports in the trackeRdashboard.
}
\keyword{datasets}
