% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{summarise_data}
\alias{summarise_data}
\title{Returns a dataframe with data summarised to the defined timebin}
\usage{
summarise_data(input_data, chemical, exp_dur, timebin, expsr_dur)
}
\arguments{
\item{input_data}{The clean data}

\item{chemical}{The name of the chemical that is going to be summarised}

\item{exp_dur}{The duration of the total experiment}

\item{timebin}{The timebin over which the data is going to be summarised}
}
\value{
A data frame that gives the mean and SD of the velocity over the timebins
}
\description{
Returns a dataframe with data summarised to the defined timebin
}
\seealso{
\code{\link{speed}, \link{accel}}
}
