% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackeRdata_summary.R
\name{summary.trackeRdata}
\alias{summary.trackeRdata}
\title{Summary of training sessions.}
\usage{
\method{summary}{trackeRdata}(object, session = NULL, movingThreshold = 0.1,
  ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be summarised, defaults to all sessions.}

\item{movingThreshold}{The threshold above which speed an athlete is considered moving.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{trackeRdataSummary}.
}
\description{
Summary of training sessions.
}
\examples{
data(runs, package = "trackeR")
runSummary <- summary(runs, session = 1:2)
## print summary
runSummary
print(runSummary, digits = 3)
## change units
changeUnits(runSummary, variable = "speed", unit = "km_per_h")
## plot summary
runSummaryFull <- summary(runs)
plot(runSummaryFull)
plot(runSummaryFull, group = c("total", "moving"),
    what = c("avgSpeed", "distance", "duration", "avgHeartRate"))
}
\seealso{
\code{\link{plot.trackeRdataSummary}}
}

