% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{test_work_capacity}
\alias{test_work_capacity}
\title{Test whether we can plot work capacity for at least one of cycling or running.}
\usage{
test_work_capacity(data)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}
}
\description{
Test whether we can plot work capacity for at least one of cycling or running.
}
