% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trackeR.R
\name{threshold}
\alias{threshold}
\title{Generic function for thresholding data.}
\usage{
threshold(object, variable, lower, upper, ...)
}
\arguments{
\item{object}{The object containing the data.}

\item{variable}{A vector containing the names of the variables to which thresholding is applied.}

\item{lower}{A vector containing the corresponding lower thresholds.}

\item{upper}{A vector containing the corresponding upper thresholds.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for thresholding data.
}

