% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thresholds.R
\name{threshold.trackeRdata}
\alias{threshold.trackeRdata}
\title{Thresholding for variables in \code{trackeRdata} objects.}
\usage{
\method{threshold}{trackeRdata}(object, variable, lower, upper, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{variable}{A vector containing the names of the variables to which thresholding is applied. See Details.}

\item{lower}{A vector containing the corresponding lower thresholds.}

\item{upper}{A vector containing the corresponding upper thresholds.}

\item{...}{Currently not used.}
}
\description{
Thresholding for variables in \code{trackeRdata} objects.
}
\details{
Argument \code{variable} can also be a data frame containing the variable names, lower, and upper thresholds.
}

