% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{timeAboveThreshold}
\alias{timeAboveThreshold}
\title{Time spent above a certain threshold}
\usage{
timeAboveThreshold(object, threshold = -1, ge = TRUE)
}
\arguments{
\item{object}{A (univariate) zoo object.}

\item{threshold}{The threshold.}

\item{ge}{Logical. Should time include the thereshold (greater or equal to threshold) or not (greater only)?}
}
\description{
Time spent above a certain threshold
}

