% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{unique.trackeRdata}
\alias{unique.trackeRdata}
\title{Exrtact unique sessions in a \code{trackerRdata} object}
\usage{
\method{unique}{trackeRdata}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{A \code{trackeRdata} object.}

\item{incomparables}{Currently not used.}

\item{...}{Currently not used.}
}
\description{
Exrtact unique sessions in a \code{trackerRdata} object
}
\details{
Uniqueness is determined by comparing the first timestamp of the
sessions in the \code{trackeRdata} object.
}
