% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{update_metrics_to_plot_selected_workouts}
\alias{update_metrics_to_plot_selected_workouts}
\title{Update metrics to plot for Work capacity and time in zones}
\usage{
update_metrics_to_plot_selected_workouts(id, session, metrics, has_data)
}
\description{
Update metrics to plot for Work capacity and time in zones
}
