% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{update_metrics_to_plot_workouts}
\alias{update_metrics_to_plot_workouts}
\title{Update panel with metrics to plot}
\usage{
update_metrics_to_plot_workouts(session, choices, data)
}
\description{
Update panel with metrics to plot
}
