% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R
\name{update_sport_selection}
\alias{update_sport_selection}
\title{Update selection of sports}
\usage{
update_sport_selection(data, session)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}

\item{session}{A shiny object.}
}
\description{
Update selection of sports
}
