% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{append_dynamics}
\alias{append_dynamics}
\title{Return speeds}
\usage{
append_dynamics(data)
}
\arguments{
\item{data}{The clean data}
}
\value{
A data frame including the speeds and accelerations
}
\description{
Return speeds
}
\seealso{
\code{\link{speed}, \link{accel}}
}
