% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrationProfile.R
\name{c2d}
\alias{c2d}
\title{Transform concentration profile to distribution profile.}
\usage{
c2d(cp)
}
\arguments{
\item{cp}{Single concentration profile as a zoo object.}
}
\description{
Transform concentration profile to distribution profile.
}
