% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{change_units.conProfile}
\alias{change_units.conProfile}
\title{Change the units of the variables in an \code{\link{conProfile}} object}
\usage{
\method{change_units}{conProfile}(object, variable, unit, ...)
}
\arguments{
\item{object}{An object of class \code{\link{conProfile}} as returned by \code{\link{concentrationProfile}}.}

\item{variable}{A vector of variables to be changed.}

\item{unit}{A vector with the units, corresponding to variable.}

\item{...}{Currently not used.}
}
\description{
Change the units of the variables in an \code{\link{conProfile}} object
}
