% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/zzz_aliases.R
\name{change_units}
\alias{change_units}
\alias{changeUnits}
\title{Generic function for changing the units of measurement}
\usage{
change_units(object, variable, unit, sport, ...)

changeUnits(object, variable, unit, sport, ...)
}
\arguments{
\item{object}{The object of which the units of measurement are changed.}

\item{variable}{A vector of variables whose units are to be changed.}

\item{unit}{A vector with the units, corresponding to \code{variable}.}

\item{sport}{A vector of sports (among \code{'cycling'},
\code{'running'}, \code{'swimming'}) with each element
corresponding to \code{variable} and \code{unit}.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for changing the units of measurement
}
