% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_helper_functions.R, R/trackeR.R,
%   R/zzz_aliases.R
\name{change_units}
\alias{change_units}
\alias{change_units}
\alias{changeUnits}
\title{Change units of variables.}
\usage{
change_units(object, variable, unit, sport, ...)

change_units(object, variable, unit, sport, ...)

changeUnits(object, variable, unit, sport, ...)
}
\arguments{
\item{object}{A character of either 'summary' or 'object' to specify which objects' units to change.}

\item{variable}{A vector of variables whose units are to be changed.}

\item{unit}{A vector with the units, corresponding to \code{variable}.}

\item{sport}{A vector of sports (amongst \code{'cycling'},
\code{'running'}, \code{'swimming'}) with each element
corresponding to \code{variable} and \code{unit}.}

\item{...}{Arguments to be passed to methods.}

\item{data}{An object of class \code{reactivevalues}.}

\item{input}{An object of class \code{reactivevalues}.}

\item{object}{The object of which the units of measurement are changed.}
}
\description{
Change units of variables.

Generic function for changing the units of measurement
}
