% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{change_units.trackeRdata}
\alias{change_units.trackeRdata}
\title{Change the units of the variables in an \code{trackeRdata} object.}
\usage{
\method{change_units}{trackeRdata}(object, variable, unit, sport, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{variable}{A vector of variables whose units are to be changed.}

\item{unit}{A vector with the units, corresponding to \code{variable}.}

\item{sport}{A vector of sports (amongst \code{'cycling'},
\code{'running'}, \code{'swimming'}) with each element
corresponding to variable and unit}

\item{...}{Arguments to be passed to methods.}
}
\description{
Change the units of the variables in an \code{trackeRdata} object.
}
