% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/zzz_aliases.R
\name{concentration_profile}
\alias{concentration_profile}
\alias{concentrationProfile}
\title{Generic method for concentration profiles}
\usage{
concentration_profile(object, session = NULL, what = NULL, ...)

concentrationProfile(object, session = NULL, what = NULL, ...)
}
\description{
Generic method for concentration profiles
}
\seealso{
concentration_profile.distrProfile
}
