% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/units.R
\name{conversions}
\alias{W2W}
\alias{W2kW}
\alias{conversions}
\alias{ft2ft}
\alias{ft2km}
\alias{ft2m}
\alias{ft2mi}
\alias{ft_per_min2ft_per_min}
\alias{ft_per_min2m_per_s}
\alias{ft_per_s2ft_per_s}
\alias{ft_per_s2m_per_s}
\alias{h2h}
\alias{h2min}
\alias{h2s}
\alias{kW2W}
\alias{kW2kW}
\alias{km2ft}
\alias{km2km}
\alias{km2m}
\alias{km2mi}
\alias{km_per_h2km_per_h}
\alias{km_per_h2m_per_s}
\alias{m2ft}
\alias{m2km}
\alias{m2m}
\alias{m2mi}
\alias{m_per_s2ft_per_min}
\alias{m_per_s2ft_per_s}
\alias{m_per_s2km_per_h}
\alias{m_per_s2m_per_s}
\alias{m_per_s2mi_per_h}
\alias{mi2ft}
\alias{mi2km}
\alias{mi2m}
\alias{mi2mi}
\alias{mi_per_h2m_per_s}
\alias{mi_per_h2mi_per_h}
\alias{min2h}
\alias{min2min}
\alias{min2s}
\alias{min_per_km2min_per_km}
\alias{min_per_km2s_per_m}
\alias{rev_per_min2rev_per_min}
\alias{s2h}
\alias{s2min}
\alias{s2s}
\alias{s_per_m2min_per_km}
\alias{s_per_m2s_per_m}
\alias{steps_per_min2steps_per_min}
\title{Auxiliary conversion functions.}
\usage{
m2km(variable)

km2m(variable)

m2ft(variable)

ft2m(variable)

m2mi(variable)

mi2m(variable)

km2ft(variable)

ft2km(variable)

km2mi(variable)

mi2km(variable)

ft2mi(variable)

mi2ft(variable)

m2m(variable)

km2km(variable)

ft2ft(variable)

mi2mi(variable)

s2min(variable)

min2s(variable)

s2h(variable)

h2s(variable)

min2h(variable)

h2min(variable)

s2s(variable)

min2min(variable)

h2h(variable)

m_per_s2km_per_h(variable)

km_per_h2m_per_s(variable)

m_per_s2ft_per_min(variable)

ft_per_min2m_per_s(variable)

m_per_s2ft_per_s(variable)

ft_per_s2m_per_s(variable)

m_per_s2mi_per_h(variable)

mi_per_h2m_per_s(variable)

m_per_s2m_per_s(variable)

km_per_h2km_per_h(variable)

ft_per_min2ft_per_min(variable)

ft_per_s2ft_per_s(variable)

mi_per_h2mi_per_h(variable)

s_per_m2min_per_km(variable)

min_per_km2s_per_m(variable)

s_per_m2s_per_m(variable)

min_per_km2min_per_km(variable)

W2kW(variable)

kW2W(variable)

W2W(variable)

kW2kW(variable)

steps_per_min2steps_per_min(variable)

rev_per_min2rev_per_min(variable)
}
\arguments{
\item{variable}{Variable to be converted.}
}
\description{
Conversion functions for distance, duration, speed, pace, power, and cadence.
}

