% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules_shiny.R
\name{create_icon}
\alias{create_icon}
\title{Generate an icon for a given feature.}
\usage{
create_icon(feature)
}
\arguments{
\item{feature}{A character representing the feature whose units we want to generate.}
}
\description{
Generate an icon for a given feature.
}
