% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_UI.R
\name{create_work_capacity_plot}
\alias{create_work_capacity_plot}
\title{Create work capacity plot}
\usage{
create_work_capacity_plot(id, cycling, collapsed = FALSE)
}
\description{
Create work capacity plot
}
