% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{fortify.conProfile}
\alias{fortify.conProfile}
\title{Fortify a \code{\link{conProfile}} object for plotting with ggplot2.}
\usage{
\method{fortify}{conProfile}(model, data, melt = FALSE, ...)
}
\arguments{
\item{model}{The \code{\link{conProfile}} object.}

\item{data}{Ignored.}

\item{melt}{Logical. Should the data be melted into long format
instead of the default wide format?}

\item{...}{Ignored.}
}
\description{
Fortify a \code{\link{conProfile}} object for plotting with ggplot2.
}
