% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/concentrationProfile.R
\name{fortify.conProfile}
\alias{fortify.conProfile}
\title{Fortify a conProfile object for plotting with ggplot2.}
\usage{
\method{fortify}{conProfile}(model, data, melt = FALSE, ...)
}
\arguments{
\item{model}{The \code{conProfile} object.}

\item{data}{Ignored.}

\item{melt}{Logical. Should the data be melted into long format
instead of the default wide format?}

\item{...}{Ignored.}
}
\description{
Fortify a conProfile object for plotting with ggplot2.
}

