% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{fortify.distrProfile}
\alias{fortify.distrProfile}
\title{Fortify a distrProfile object for plotting with ggplot2.}
\usage{
\method{fortify}{distrProfile}(model, data, melt = FALSE, ...)
}
\arguments{
\item{model}{The \code{distrProfile} object.}

\item{data}{Ignored.}

\item{melt}{Logical. Should the data be melted into long format
instead of the default wide format?}

\item{...}{Ignored.}
}
\description{
Fortify a distrProfile object for plotting with ggplot2.
}

