% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_output.R
\name{generate_selected_sessions_object}
\alias{generate_selected_sessions_object}
\title{Generate an object with selected sessions}
\usage{
generate_selected_sessions_object(data, input)
}
\arguments{
\item{data}{An object of class \code{reactivevalues}.}

\item{input}{A shiny object with user input.}
}
\description{
Generate an object with selected sessions
}
