% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R, R/zzz_aliases.R
\name{generate_thresholds}
\alias{generate_thresholds}
\alias{generateDefaultThresholds}
\title{Generate default thresholds.}
\usage{
generate_thresholds(variable, lower, upper, sport, ...)

generateDefaultThresholds(variable, lower, upper, sport, ...)
}
\arguments{
\item{variable}{A vector of variables with user-specified thresholds.}

\item{lower}{A vector of lower limits corresponding to the elements of \code{variable}.}

\item{upper}{A vector of upper limits corresponding to the elements of \code{variable}.}

\item{sport}{A vector of sports (amongst \code{'cycling'},
\code{'running'}, \code{'swimming'}) with each element
corresponding to \code{variable}, \code{lower} and \code{upper}.}

\item{...}{Currently not used.}
}
\description{
Generate default thresholds.
}
