% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/units.R
\name{generateBaseUnits}
\alias{generateBaseUnits}
\title{Generate base units.}
\usage{
generateBaseUnits(cycling = FALSE, ...)
}
\arguments{
\item{cycling}{Logical. Is the data from a cycling session rather than a running session?}

\item{...}{Currently not used.}
}
\description{
Generate base units.
}

