% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R, R/trackeR.R
\name{get_profile.distrProfile}
\alias{get_profile.distrProfile}
\alias{get_profile.conProfile}
\alias{get_profile}
\title{Generic function to subset distribution and concentration profiles}
\usage{
\method{get_profile}{distrProfile}(object, session = NULL, what = NULL,
  ...)

\method{get_profile}{conProfile}(object, session = NULL, what = NULL,
  ...)

get_profile(object, session, what, ...)
}
\arguments{
\item{object}{An object of class \code{distrProfile} or \code{conProfile}
as returned by \code{\link{distribution_profile}} and
\code{\link{concentration_profile}}, respectively.}

\item{session}{A numeric vector of the sessions to selected.
Defaults to all sessions.}

\item{what}{A character version of the variables to be
selected. Defaults to all variables in \code{object}
(\code{what = NULL}).}

\item{...}{Current no used.}
}
\description{
Generic function to subset distribution and concentration profiles
}
