% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wprime.R, R/distributionProfile.R, R/trackeR.R,
%   R/trackeRdata.R, R/trackeRdata_summary.R
\name{get_sport.trackeRWprime}
\alias{get_sport.trackeRWprime}
\alias{get_sport.distrProfile}
\alias{get_sport}
\alias{get_sport.trackeRdata}
\alias{get_sport.trackeRdataSummary}
\title{Generic function for extracting sports}
\usage{
\method{get_sport}{trackeRWprime}(object, ...)

\method{get_sport}{distrProfile}(object, session = NULL, ...)

get_sport(object, session, ...)

\method{get_sport}{trackeRdata}(object, session = NULL, ...)

\method{get_sport}{trackeRdataSummary}(object, session = NULL, ...)
}
\arguments{
\item{object}{The object from which to extract sports.}

\item{...}{Arguments to be passed to methods.}

\item{session}{The sessions for which to extract sports.}
}
\description{
Generic function for extracting sports
}
