% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wprime.R, R/distributionProfile.R, R/trackeR.R,
%   R/trackeRdata_summary.R
\name{nsessions.trackeRWprime}
\alias{nsessions.trackeRWprime}
\alias{nsessions.distrProfile}
\alias{nsessions.conProfile}
\alias{nsessions}
\alias{nsessions.trackeRdataSummary}
\title{Generic function for calculating number of sessions}
\usage{
\method{nsessions}{trackeRWprime}(object, ...)

\method{nsessions}{distrProfile}(object, ...)

\method{nsessions}{conProfile}(object, ...)

nsessions(object, ...)

\method{nsessions}{trackeRdataSummary}(object, ...)
}
\arguments{
\item{object}{The object for which to calculate the number of sessions.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for calculating number of sessions
}
