% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{nsessions}
\alias{nsessions}
\title{Generic function for calculating number of sessions}
\usage{
nsessions(object, variable, unit, ...)
}
\arguments{
\item{object}{An object of class trackeRdata, trackeRdataSummary, trackeRdataZones, conProfile, or distrProfile}
}
\description{
Generic function for calculating number of sessions
}

