% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata_summary.R
\name{plot.trackeRdataSummary}
\alias{plot.trackeRdataSummary}
\title{Plot an object of class trackeRdataSummary.}
\usage{
\method{plot}{trackeRdataSummary}(x, date = TRUE, what = NULL,
  group = NULL, lines = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{trackeRdataSummary}.}

\item{date}{Should the date or the session number be used on the abscissa?}

\item{what}{Name of variables which should be plotted. Default is all.}

\item{group}{Which group of variables should be plotted? This can either be
\code{total} or \code{moving}. Default is both.}

\item{lines}{Should interpolating lines be plotted?}

\item{...}{Currently not used.}
}
\description{
Plot an object of class trackeRdataSummary.
}
\examples{
data("runs", package = "trackeR")
runSummary <- summary(runs)
plot(runSummary)
plot(runSummary, date = FALSE, group = "total",
    what = c("distance", "duration", "avgSpeed"))
}
\seealso{
\code{\link{summary.trackeRdata}}
}
