% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plotSelectedWorkouts.R
\name{plot_selectedWorkouts}
\alias{plot_selectedWorkouts}
\title{Plot the profile for each selected session for a given variable (heart rate, altitude, pace).}
\usage{
plot_selectedWorkouts(x, session, what, sumX, threshold = TRUE,
  smooth = FALSE, trend = TRUE, dates = TRUE, plotly = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be plotted, defaults to all sessions.}

\item{what}{Which variables should be plotted?}

\item{sumX}{An object of class \code{\link{trackeRdataSummary}}.}

\item{threshold}{Logical. Should thresholds be applied?}

\item{smooth}{Logical. Should the data be smoothed?}

\item{trend}{Logical. Should a smooth trend be plotted?}

\item{dates}{Logical. Should the date of the session be used in the panel header?}

\item{plotly}{Logical. Whether return plotly plots or standard TrackeR plot.}
}
\description{
Plot the profile for each selected session for a given variable (heart rate, altitude, pace).
}
