% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plotSelectedWorkouts.R
\name{plot_selectedWorkouts}
\alias{plot_selectedWorkouts}
\title{Plot the profile for each selected session for a given variable (heart rate, altitude, pace).}
\usage{
plot_selectedWorkouts(x, session, what, sumX, threshold = TRUE,
  smooth = FALSE, trend = TRUE, dates = TRUE, changepoints = FALSE,
  n_changepoints = 6, print_changepoints = FALSE)
}
\arguments{
\item{x}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be plotted, defaults to all sessions.}

\item{what}{Which variables should be plotted?}

\item{sumX}{An object of class \code{trackeRdataSummary}.}

\item{threshold}{Logical. Should thresholds be applied?}

\item{smooth}{Logical. Should the data be smoothed?}

\item{trend}{Logical. Should a smooth trend be plotted?}

\item{dates}{Logical. Should the date of the session be used in the panel header?}

\item{changepoints}{Logical. Whether changepoints should be identified and plotted.}

\item{n_changepoints}{A numeric. The threshold for the maximum number of changepoints to search for.}

\item{print_changepoints}{Logical. Whether or not to print changepoint values (when changepoints = TRUE).}
}
\description{
Plot the profile for each selected session for a given variable (heart rate, altitude, pace).
}
