% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_work_capacity.R
\name{plot_work_capacity}
\alias{plot_work_capacity}
\title{Plot the work capacity W' (w prime).}
\usage{
plot_work_capacity(x, session, plotly = TRUE, dates = TRUE, scaled = TRUE)
}
\arguments{
\item{x}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be used, defaults to all sessions.}

\item{plotly}{A boolean as to whether an output should be a plotly plot or a standard TrackeR plot.}

\item{dates}{Logical. Should the date of the session be used in the panel header?}

\item{scaled}{Logical. Should the W' be scaled to the movement variable (power or speed)
which is then plotted in the background?}
}
\description{
Plot the work capacity W' (w prime).
}
