% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_workouts.R
\name{plot_workouts}
\alias{plot_workouts}
\title{Plot an object of class trackeRdataSummary.}
\usage{
plot_workouts(sumX, what, sessions, shiny = TRUE, date = TRUE,
  group = c("total"), lines = TRUE, sports)
}
\arguments{
\item{sumX}{An object of class \code{trackeRdataSummary}.}

\item{what}{Name of variables which should be plotted. Default is all.}

\item{sessions}{A vector. Selected sessions by session number.}

\item{shiny}{Logical. Whether plots are in a shiny environment.}

\item{date}{Should the date or the session number be used on the abscissa?}

\item{group}{Which group of variables should be plotted? This can either be
\code{total} or \code{moving}. Default is both.}

\item{lines}{Should interpolating lines be plotted?}

\item{sports}{A vector of sports of the sessions to be plotted.}

\item{...}{Currently not used.}
}
\description{
Plot an object of class trackeRdataSummary.
}
\seealso{
\code{\link{summary.trackeRdata}}
}
