% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_zones.R
\name{plot_zones}
\alias{plot_zones}
\title{Plot training zones}
\usage{
plot_zones(run_data, session, what = c("heart.rate"))
}
\arguments{
\item{run_data}{An object of class \code{trackeRdata} as returned by \code{\link{readDirectory_shiny}}.}

\item{session}{A vector of selected sessions.}

\item{what}{A vector of variable names to be plotted.}
}
\description{
Plot training zones
}
