% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_zones.R
\name{plot_zones}
\alias{plot_zones}
\title{Plot training zones.}
\usage{
plot_zones(x, session, what = c("heart_rate"), n_zones)
}
\arguments{
\item{x}{An object of class \code{trackeRdata}.}

\item{session}{A vector of selected sessions.}

\item{what}{A vector of variable names to be plotted.}

\item{n_zones}{A numeric. The number of zones to split the dataset into.}
}
\description{
Plot training zones.
}
