% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{print.trackeRdata}
\alias{print.trackeRdata}
\title{\code{\link{print}} method for \code{\link{trackeRdata}} objects}
\usage{
\method{print}{trackeRdata}(x, duration = "h", ..., digits = 2)
}
\arguments{
\item{x}{An object of class \code{\link{trackeRdata}}.}

\item{...}{Currently not used; only for compatibility with generic \code{\link{summary}} method only.}

\item{digits}{Number of digits to be printed.}
}
\description{
\code{\link{print}} method for \code{\link{trackeRdata}} objects
}
\details{
The print method returns training coverage, number of sessions and
total training duration from the data in the
\code{\link{trackeRdata}} object.
}
