% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{print.trackeRdata}
\alias{print.trackeRdata}
\title{print method for \code{\link{trackeRdata}} objects}
\usage{
\method{print}{trackeRdata}(x, duration = "h", ..., digits = 2)
}
\arguments{
\item{x}{an object of class \code{\link{trackeRdata}}}

\item{...}{currently not used; only for compatibility with generic \code{\link{summary}} method only}

\item{digits}{number of digits to be printed}
}
\description{
print method for \code{\link{trackeRdata}} objects
}
\details{
The print method returns training coverage, number of sessions and
total training duration from the data in the
\code{\link{trackeRdata}} object.
}
