% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{run}
\alias{run}
\title{Training session.}
\format{A \code{\link{trackeRdata}} object containing one running training session.}
\usage{
run
}
\description{
Training session.
}
\keyword{datasets}

