% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{scaled.distrProfile}
\alias{scaled.distrProfile}
\title{Scale the distribution profile relative to its maximum value}
\usage{
\method{scaled}{distrProfile}(x, session = NULL, what = c("speed",
  "heart.rate"))
}
\arguments{
\item{session}{A numeric vector of the sessions to be plotted, defaults to all sessions.}

\item{what}{Which variables should be scaled?}

\item{object}{An object of class \code{distrProfile} as returned by \code{\link{distributionProfile}}.}
}
\description{
Scale the distribution profile relative to its maximum value
}

