% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\name{scaled}
\alias{scaled}
\title{Generic function for scaling.}
\usage{
scaled(object, ...)
}
\arguments{
\item{object}{The object to be scaled.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for scaling.
}
