% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R, R/trackeRdata.R
\name{session_duration}
\alias{session_duration}
\alias{session_duration.trackeRdata}
\title{Generic function for calculating session durations.}
\usage{
session_duration(object, ...)

\method{session_duration}{trackeRdata}(object, ...)
}
\arguments{
\item{object}{The object for which to calculate session durations.}

\item{...}{Arguments to be passed to methods.}
}
\description{
Generic function for calculating session durations.
}
\details{
The times units will be inherited from \code{object}.
}
