% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp_plot_map.R
\name{shiny_plot_map}
\alias{shiny_plot_map}
\title{Plot all selected workouts on an interactive leaflet map.}
\usage{
shiny_plot_map(x, sumX, session = NULL, threshold = TRUE,
  shiny_version = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{\link{trackeRdata}}.}

\item{sumX}{An object of class \code{\link{trackeRdataSummary}}.}

\item{session}{A numeric vector of the sessions to be plotted. Defaults to
all sessions.}

\item{threshold}{Logical. Should thresholds be applied?}

\item{shiny_version}{Logical. Whether plot shiny version of map or TrackeR version.}
}
\description{
Plot all selected workouts on an interactive leaflet map.
}
