% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionProfile.R
\name{smootherControl.distrProfile}
\alias{smootherControl.distrProfile}
\title{Auxiliary function for \code{\link{smoother.distrProfile}}. Typically used to construct
a control argument for \code{\link{smoother.distrProfile}}.}
\usage{
smootherControl.distrProfile(what = c("speed", "heart.rate"), k = 30,
  sp = NULL, parallel = FALSE, cores = NULL)
}
\arguments{
\item{what}{Vector of the names of the variables which should be smoothed.}

\item{k}{Number of knots.}

\item{sp}{A vector of smoothing parameters passed on to \code{\link[scam]{scam}}.}

\item{parallel}{Logical. Should computation be carried out in parallel?}

\item{cores}{Number of cores for parallel computing. If NULL, the number of cores is set to the value of \code{options("corese")} (on Windows) or \code{options("mc.cores")} (elsewhere), or, if the relevant option is unspecified, to half the number of cores detected.}
}
\description{
Auxiliary function for \code{\link{smoother.distrProfile}}. Typically used to construct
a control argument for \code{\link{smoother.distrProfile}}.
}

