% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{sort.trackeRdata}
\alias{sort.trackeRdata}
\title{Sort the sessions \code{trackeRdata} objects into ascending or
descending order according to the first session timestamp.}
\usage{
\method{sort}{trackeRdata}(x, decreasing = FALSE)
}
\description{
Sort the sessions \code{trackeRdata} objects into ascending or
descending order according to the first session timestamp.
}
