% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/trackeRdata_summary.R
\name{summary.trackeRdata}
\alias{summary.trackeRdata}
\title{Summary of training sessions.}
\usage{
\method{summary}{trackeRdata}(object, session = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{\link{trackeRdata}}.}

\item{session}{A numeric vector of the sessions to be summarised, defaults to all sessions.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{trackeRdataSummary}.
}
\description{
Summary of training sessions.
}
\examples{
data(run, package = "trackeR")
runSummary <- summary(run)
runSummary
print(runSummary, digits = 3)
plot(runSummary)
}
\seealso{
\code{\link{plot.trackeRdataSummary}}
}

