% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\docType{package}
\name{trackeR}
\alias{trackeR}
\alias{trackeR-package}
\title{trackeR: Infrastructure for for handling cycling and running data from GPS-enabled tracking devices}
\description{
trackeR provides infrastructure for handling cycling and running
data from GPS-enabled tracking devices. After extraction and appropriate
manipulation of the training or competition attributes, the data are placed into
session-aware data objects with an S3 class trackeRdata. The information in the
resultant data objects can then be visualised, summarised and analysed through
corresponding flexible and extensible methods.
}

