% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeR.R
\docType{package}
\name{trackeR}
\alias{trackeR}
\alias{trackeR-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

