% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRdata.R
\name{unique.trackeRdata}
\alias{unique.trackeRdata}
\title{Exrtact unique sessions in a \code{trackerRdata} object.}
\usage{
\method{unique}{trackeRdata}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{a \code{trackeRdata} object}

\item{incomparables}{currently not used}

\item{...}{currently not used}
}
\description{
Exrtact unique sessions in a \code{trackerRdata} object.
}
\details{
Uniqueness is determined by comparing the first timestamp of the
sessions in the \code{trackeRdata} object.
}
